+++ src/main/java/com/beotkkotthon/areyousleeping/domain/specification/BoardSpecifications.java
@@ -1,0 +1,2 @@
+package com.beotkkotthon.areyousleeping.domain.specification;public class BoardSpecifications {
+}
